# [Exec](expr) Opt the compound pred performace (#45414)